refactor(annotation): extract NiceGUI-free functional core
annotation_core.py carries tag/highlight/item serialisation with no
NiceGUI import (guarded by a fresh-interpreter subprocess test) so the
standalone snapshot service can reuse it. tags.py and items_serialise.py
become re-export shims; card_shared and highlights consume the core.
